-removed 'credentials model' - fixed role type - fixed package.json dependencies - re-installed bcryptjs, cookie-parser -user model: added the timestamp option. -properly re-named folder names (removed numbering)
