+++ server/src/controllers/user.controller.ts
@@ -1,0 +1,1 @@
+import UserModel from "../models/user.model";
